Ongoing work on the incoming set
